👌 Truncate client log to 3000 bytes
