+++ pkg/middleware/logger/logger_test.go
@@ -76,37 +76,3 @@ func TestLogging(t *testing.T) {
 		})
 	}
 }
-
-func TestTruncate(t *testing.T) {
-	testCases := []struct {
-		name      string
-		input     string
-		maxLength int
-		expected  string
-	}{
-		{
-			name:      "Should truncate string",
-			input:     "1234567890",
-			maxLength: 5,
-			expected:  "12345",
-		},
-		{
-			name:      "Should keep string if it's equal to maxLength",
-			input:     "1234567890",
-			maxLength: 10,
-			expected:  "1234567890",
-		},
-		{
-			name:      "Should keep string if it's over maxLength",
-			input:     "1234567890",
-			maxLength: 11,
-			expected:  "1234567890",
-		},
-	}
-	for _, tc := range testCases {
-		tc := tc
-		t.Run(tc.name, func(t *testing.T) {
-			assert.Equal(t, tc.expected, logger.Truncate(&tc.input, tc.maxLength))
-		})
-	}
-}
